+++ test/libsolidity/syntaxTests/inheritance/interface/overrides_single.sol
@@ -11,4 +11,3 @@ interface Sub is Super {
 }
 
 // ----
-// TypeError 9456: (197-241): Overriding function is missing "override" specifier.