+++ test/libsolidity/syntaxTests/inheritance/override/diamond_interface_intermediate_public_state_variable_and_function.sol
@@ -3,12 +3,12 @@ interface I {
 }
 contract A is I
 {
-	uint public override f;
+	uint public f;
 }
 abstract contract B is I
 {
-	function f() external virtual override returns (uint);
+	function f() external virtual returns (uint);
 }
 abstract contract C is A, B {}
 // ----
-// TypeError 6480: (185-215): Derived contract must override function "f". Two or more base classes define function with same name and parameter types. Since one of the bases defines a public state variable which cannot be overridden, you have to change the inheritance layout or the names of the functions.
+// TypeError 6480: (167-197): Derived contract must override function "f". Two or more base classes define function with same name and parameter types. Since one of the bases defines a public state variable which cannot be overridden, you have to change the inheritance layout or the names of the functions.