+++ test/libsolidity/syntaxTests/inheritance/override/implement_interface_by_public_variable.sol
@@ -1,6 +1,6 @@
 interface X { function test() external returns (uint256); }
 contract Y is X {
-    uint256 public override test = 42;
+    uint256 public test = 42;
 }
 contract T {
     constructor() { new Y(); }